developing scene for wheel roller
got uid error but i believe it was just the project getting confused because I deleted an early scene prototype. Put the uid in the "Keep in Mind" section of `TODO.md` just in case. I was able to replicate it by clicking "Quick Open" and fixed it by deleting the folder that the scene was held in.